+++ src/main/app.ts
@@ -30,6 +30,7 @@ import {
   MenuThemeDescription,
   WMInfo,
   SoundType,
+  FileIconThemeDescription,
   SoundEffect,
   CommandlineOptions,
 } from '../common';
@@ -253,6 +254,11 @@ export class KandoApp {
       return true;
     }
 
+    if (options.reloadIconThemes) {
+      this.reloadIconThemes();
+      return true;
+    }
+
     return false;
   }
 
@@ -386,6 +392,17 @@ export class KandoApp {
   }
 
   /**
+   * This is called when the --reload-icon-themes command line option is passed or when
+   * the respective button in the settings is pressed.
+   */
+  public async reloadIconThemes() {
+    delete this.iconThemesCache;
+
+    this.menuWindow?.webContents.send('menu-window.reload-icon-themes');
+    this.settingsWindow?.webContents.send('settings-window.reload-icon-themes');
+  }
+
+  /**
    * Setup IPC communication with the renderer process. See ../renderer/preload.ts for
    * more information on the exposed functionality.
    */
@@ -516,6 +533,11 @@ export class KandoApp {
     // Reload the current sound theme if requested.
     ipcMain.on('settings-window.reload-sound-theme', async () => {
       this.reloadSoundTheme();
+    });
+
+    // Reload the current icon themes if requested.
+    ipcMain.on('settings-window.reload-icon-themes', async () => {
+      this.reloadIconThemes();
     });
 
     // Allow showing open-file dialogs.
@@ -730,28 +752,30 @@ export class KandoApp {
         return this.iconThemesCache;
       }
 
-      this.iconThemesCache = {
-        userIconDirectory: path.join(app.getPath('userData'), 'icon-themes'),
-        fileIconThemes: [],
-      };
-
       const themes = await this.listSubdirectories([
         path.join(app.getPath('userData'), 'icon-themes'),
         path.join(__dirname, '../renderer/assets/icon-themes'),
       ]);
+
+      const fileIconThemes: FileIconThemeDescription[] = [];
 
       await Promise.all(
         themes.map(async (theme) => {
           const directory = await this.findIconThemePath(theme);
           const icons = await this.listIconsRecursively(directory);
 
-          this.iconThemesCache.fileIconThemes.push({
+          fileIconThemes.push({
             name: theme,
             directory,
             icons,
           });
         })
       );
+
+      this.iconThemesCache = {
+        userIconDirectory: path.join(app.getPath('userData'), 'icon-themes'),
+        fileIconThemes,
+      };
 
       return this.iconThemesCache;
     });