+++ docs/changelog.md
@@ -24,6 +24,8 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 ### :tada: Added
 
 - Buttons to the general settings dialog to **backup and restore** the `config.json` and `menus.json` files. This makes it easier to backup your settings before making big changes or to restore them from a previous backup.
+- A button in the icon-picker popover to **reload all icon themes** without restarting Kando.
+- The command line flag `--reload-icon-themes` which can be used to tell a running Kando instance to reload all icon themes.
 - A setting to **enable or disable the drawing of selection wedges** in supported menu themes. This can be found in the theme-settings dialog. It is disabled by default as the selection wedges are quite resource-intensive to draw currently.
 
 ### :wrench: Changed